# Use `spawn_batch` in `many_lights` example